fix: implemented review feedback
